pyproject.toml
[tool.poetry]
name = "tickapp"
version = "0.1.0"
description = ""
authors = ["mileto <carmeloenzomileto@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.14"
dagster = "^1.12.3"
python-dotenv = "^1.2.1"
dagster-webserver = "^1.12.3"
anthropic = "^0.74.1"
sqlalchemy = "^2.0.44"
alembic = "^1.17.2"
psycopg2 = "^2.9.11"
dagster-postgres = "^0.28.3"

[tool.dagster]
module_name = "tickapp.assets"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
